lambda/requirements.txt
python-dotenv
requests
dnspython
slackclient
